+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 - Added list program account receipt endpoint
 - Added list user receipt endpoint
 - Added list prepaid card receipt endpoint
+- Added list program account balance endpoint
 
 0.1.1 (2016-06-30)
 ------------------